requirements.txt
pexpect>=4.2.1
pyyaml